fix: tablebase WDL probing for nonzero halfmove clocks + correctness hardening
probeWDL() was returning UNKNOWN as soon as halfmoveClock != 0, blinding
subtree probing after any reversible move. Now falls back to
tb_probe_root() + TB_GET_WDL() for nonzero rule-50 counts while keeping
the fast tb_probe_wdl() path at 0.

Also fixes signed integer overflow in AI.cpp (isPVNode, null-move eval),
adds UBSan-safe null guards to Bitboard lsb/msb/popLsb, hardens
Polyglot book loading (file-size validation, atomic swap, uint64
weights, equal_range), silences logger in UCI mode, and adds a CMake
sanitizer option. Test coverage expanded across all changes.

+++ src/AI.cpp
@@ -414,7 +414,7 @@ int AI::negamax(Position& pos, int depth, int alpha, int beta, int ply, Move exc
     return *score;
   }
 
-  bool isPVNode = (beta - alpha) > 1;
+  bool isPVNode = (static_cast<long long>(beta) - static_cast<long long>(alpha)) > 1;
 
   // [Improvement 10] Fix: futility pruning NOT applied at PV nodes
   auto pruning = canPrune(pos, depth, alpha, beta, isPVNode);
@@ -826,7 +826,7 @@ std::optional<int> AI::tryNullMovePruning(Position& pos, int depth, int /*alpha*
 
     // Eval-based boost: if static eval is well above beta, increase R
     int eval = Eval::evaluate(pos);
-    if (eval > beta + 200) R++;
+    if (static_cast<long long>(eval) > static_cast<long long>(beta) + 200) R++;
 
     R = std::min(R, depth - 1);  // Don't reduce below depth 1
